% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batter_viz_func.R
\name{add_numbers}
\alias{add_numbers}
\title{Get dataframe of individual players' wRC by date}
\usage{
add_numbers(team_top, yr)
}
\arguments{
\item{team_top}{a vector of player_id values}

\item{y}{A numeric value representing the four digit year}
}
\value{
A tibble with daily wRC for multiple players
}
\description{
Get dataframe of individual players' wRC by date
}
