% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batter_viz_func.R
\name{complete_wrc_df}
\alias{complete_wrc_df}
\title{Complete and clean a dataframe for visualizing}
\usage{
complete_wrc_df(tm_wrc_by_date)
}
\arguments{
\item{tm_wrc_by_date}{tibble with players and wRC by date}
}
\value{
a tibble with all date and player combinations
}
\description{
Complete and clean a dataframe for visualizing
}
