% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batter_viz_func.R
\name{generate_viz}
\alias{generate_viz}
\title{Generate the offensive wRC plot by player over a season}
\usage{
generate_viz(tm, yr)
}
\arguments{
\item{tm}{string representing MLB team using fangraphs standards}

\item{yr}{four digit numeric year}
}
\value{
a tibble with all date and player combinations
}
\description{
Generate the offensive wRC plot by player over a season
}
