% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batter_viz_func.R
\name{get_daily_wrcs}
\alias{get_daily_wrcs}
\title{Get dataframe of individual players' wRC by date}
\usage{
get_daily_wrcs(team_top, yr)
}
\arguments{
\item{team_top}{tibble with player_ids}

\item{y}{A numeric value representing the four digit year}
}
\value{
A tibble with daily wRC for multiple players
}
\description{
Get dataframe of individual players' wRC by date
}
