% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batter_viz_func.R
\name{get_top_batters}
\alias{get_top_batters}
\title{This function gets the FanGraphs Player IDs for the 
top n players for a given team and year.}
\usage{
get_top_batters(tm, yr)
}
\arguments{
\item{tm}{Character value indicating the team.}

\item{yr}{A numeric value indicating the four digit year.}
}
\value{
A character vector of player_ids
}
\description{
This function gets the FanGraphs Player IDs for the 
top n players for a given team and year.
}
