% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batter_viz_func.R
\name{viz_wrc}
\alias{viz_wrc}
\title{Build a visualization}
\usage{
viz_wrc(complete_df, tm, yr, cp)
}
\arguments{
\item{tm_wrc_by_date}{tibble with players and wRC by date}
}
\value{
a tibble with all date and player combinations
}
\description{
Build a visualization
}
