% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batter_viz_func.R
\name{get_player_daily_wrc}
\alias{get_player_daily_wrc}
\title{This function gets the daily wRC for a given batter for a given year.}
\usage{
get_player_daily_wrc(player_id, yr)
}
\arguments{
\item{player_id}{A character value representing a fangraphs playerid.}

\item{yr}{A numeric value indicating four digit year.}
}
\value{
A tibble containing daily wRC for a season
}
\description{
This function gets the daily wRC for a given batter for a given year.
}
